Lazy type registry
Summary:
All Thrift generated classes (struct, exception, union) are loaded when conformance test server is started. This is taking in worst case 3-4sec.
The goal here is to create a list of mapping (uri = className) and register objects to type registry lazy.
Type registry is checking this mapping list when uri/hash is not found in the registry and add it to the registry.
Mapping list is stored in the new unique file names _TypeList_<unique hash code>.
With this change, registering objects in conformance test dropped ~500ms. This includes scanning the classpath to find all classes with name "_TypeList_*"

t_mstch_java2_generator.cc needs some improvements. (We should pass a list to the mustache file instead of string etc).

Differential Revision: D36810629

fbshipit-source-id: fd9512d9e797ef66f1e8fd87cedb3b89004cbe5f