Add __fbthrift_clear_terse_fields
Summary:
All terse fields need to be cleared before deserialization for consistent behavior. If a field has a custom default we don't know whether the serialization was skipped for the field since it was empty or IDL version mismatch. Therefore, to have consistent behavior with terse fields with custom default and no custom field, we need to clear all terse fields regardless with custom default.

Create __fbthrift_clear_terse_fields iff a struct has a terse field. This will be called before deserialization of the struct.

Reviewed By: Mizuchi

Differential Revision: D36957304

fbshipit-source-id: 482fe0aa8d96261ba119a28e775ef23d7d4232ca