Allowing WildcardMethodMetadata to specify executor type
Summary:
Some wildcard processors are executing requests in-line by nature. In order to support ResourcePool for those processor, we need the processor to carry the information of executing inline to the server/transport.
WildcardMethodMetadata is the best place to achieve it.

The following diff will supply the info on whether to enable ResourcePool

Reviewed By: praihan

Differential Revision: D36984332

fbshipit-source-id: 10399ab87cbadd6a685c7ba636629c441502e89c