Remove proxy exception header translation from Rocket and enable passing proxiedPayloadMetadata instead
Summary: This simplifies both rocket code and proxy code. The HTTP2 code has to be updated to do the translation to avoid breaking the wire format. This change breaks the wire format for proxied exceptions over Header transport - but header transport should never be used because of the Header to Rocket auto upgrade.

Reviewed By: iahs

Differential Revision: D36887427

fbshipit-source-id: b3ced6cce967aa408a1b5d6cba40a72f9c1b5351