Pass TransportInfo to canAcceptConnection
Summary: Transport info contains information like the security properties of the connection.  Having that passed into canAcceptConnection could be useful if a custom routing handler doesn't want to support an entire alpn but instead a subset of the requests within it.  For instance, don't take over all of http/1.1 requests but only those that might be for CONNECT or some other method.  At the same time, the handler can ensure the transport is secure.

Reviewed By: andriigrynenko

Differential Revision: D37315523

fbshipit-source-id: 50dd34fd9a46a55303da9b6486488b9c37110a8e