Add optional FlushPolicy to FlushManager.
Summary:
Currently there's two ways RocketClients have their pending writes flushed. One is by the externally managed FlushList. This is useful for direct users of RocketClient. The default policy is to batch flushes in an event loop and flush all of them in the next loop callback. This is a good general policy.

However, there are cases where this amount of flushing is excessive and it is not feasible to directly control the flushing via a FlushList. For example, if the RocketClients are pooled in a library proxy. For such usecases it can still be useful for the service to control the general flushing behavior.

This adds the ability to set a flush policy consisting of two parameters: the max number of flushes and a maximum latency. If the max number of flushes is reached, the flush will occur in the next loop callback. If the limit is never reached, an AsyncTimeout at the given max latency will trigger the flush.

The policy is set per-EventBase, as this is how the existing FlushManager is utilized.

Reviewed By: kvtsoy

Differential Revision: D37085134

fbshipit-source-id: c694df74a6284290599f48fe6cfb0a36a9c680f1