Use generated type const in codegen wherever typedef is used based on the compiler option "typedef"
Summary:
Facebook :

We generate Hack type const for typedef when `typedef` compiler option is used. But we don't use it anywhere. Wherever the typedef is used, we resolve it to original type

Consider this eg

```
Foo.thrift
    include "Bar.thriift"

    struct MyStruct {
        1: Bar.typeconst field;
    }
```

```
Bar.thrift
    typedef i32 typeconst;
```

In generated code, type of  `field` in `MyStruct` is `int` and not `typeconst`.

This diff fixes that.

Both `Foo` and `Bar` need to have `typedef` option. If only Foo uses that option it will cause errors. Since `typeconst` will not be defined in the generated code for Bar.

#forcetdhashing

Reviewed By: prasad223

Differential Revision: D37152205

fbshipit-source-id: 783b8c1be37d8654ac5d02ca987bfcfc22440d81