More test cases in fixture for streaming
Summary:
We have two modifications here:

We now have different names for exceptions thrown by stream and service. This is to check if they're handled correctly in the generated code. Right now, it's difficult to distinguish them while reading the generated code.

Also added some more test cases with different combinations of stream and service throws. These test cases will be relevant in a following diff where we'll implement streaming method in server side.

Side note: I am not sure how these fixtures are tested. I made some faulty changes in `server.mustache` which would break the build in my own test thrift files, but the code would just generate fine in the textures. Not sure if smth would break in the phabricator here in that case or not.

Reviewed By: wqfish

Differential Revision: D37445205

fbshipit-source-id: ad4d340686e491a92e76a34fb997193d97abbdc0