Implement the ability to get field mask by field id
Summary:
Defined MaskRef's `MaskRef get(FieldId id);` method. This gets the inner field mask from a mask for the given field id. MaskRef is used for the input and output of this function.
MaskRef struct represents the Field Mask and whether the mask is coming from exclusive mask or not.
```
struct MaskRef {
  const Mask& mask;
  bool is_exclusion = false;
};
```
MaskRef is used to determine the status of the mask, because if the mask is coming from exclusive mask, the mask actually represents the complement set.
This method can be chained (e.g. `MaskRef{mask, false}.get(9).get(5)` to get the inner field mask.
Also, `bool isAllMask()` and `bool isNoneMask()` methods are added to MaskRef struct to help with this get method. They return true if the given MaskRef represents allMask/ noneMask.

Reviewed By: iahs

Differential Revision: D37464049

fbshipit-source-id: 4f6b30a48f08da25c130ee4575dc4f5a17e40fae