Add ReservedIds annotation and validator to disallow specific field ids
Summary:
This adds a new annotation to disallow field ids from being used. This can be for multiple reasons: fields are deprecated, or the fields need to be reserved for some other use case. This annotation can be applied on structs, unions, exceptions, and enums, and it will be a build failure if a disallowed id is used.

For example, it is very common to see comments in thrift files like this:

```
struct Foo {
  // DEPRECATED: 1
  2: i32 a;
}
```
but nothing actually enforces that field 1 is not reused. Instead, we can annotate the struct with this:

```
include "thrift/annotation/thrift.thrift"

thrift.ReservedIds{ids = [1]}
struct Foo {
  // Build failure if 1 is used
  2: i32 a;
}
```

If tag value 1 is reused, then this will be a compilation failure. The annotation also allows providing id ranges if many values are deprecated:

```
thrift.ReservedIds{id_ranges = {10: 15, 20: 30}}
struct Foo {
  // Half-open ranges [10, 15) and [20, 30) are disallowed
  ...
}
```

Reviewed By: Mizuchi, iahs

Differential Revision: D37480975

fbshipit-source-id: 8514d64c33d990d0b875eef9003566df6ee7caba