Implement clear(...) to clear masked fields in schemaless thrift object
Summary:
Implemented `void clear(const Mask& mask, protocol::Object& obj)` to clear masked fields in schemaless thrift object (a.k.a) Protocol Object. It goes through the obj recurisvely and erases the fields that are masked. It throws a runtime exception when the field mask and object are incompatible, meaning the mask is for object type (not noneMask or allMask) but the field is non-object type.
Added void MaskRef::clear(protocol::Object& obj) method that does the same thing but with MaskRef. This allows calling the function recursively as it knows if the inner mask is coming from exclusive mask or not.

Reviewed By: thedavekwon

Differential Revision: D37494738

fbshipit-source-id: 00113b352d0964b88d6e591df8c4e9ca83053e17