Implement optimized 64-bit varint decoding functions.
Summary:
Implements optimized decoding of 64-bit varints using the BMI2 pext instruction
(when available) or a shift/shuffle/mask SSE strategy when not. Microbenchmarks
show that the new versions are substantially faster for small (1-2 byte
encodings) or large (6-10 byte encodings) varints, while being similar for the
middle range. This is the sort of performance profile we want given the
emprical distributions of varints; they tend to be heavily biased towards the
very small or very large parts of their range.

I think the microbenchmarks likely under-sell the improvement, though. The
current (unrolled) implementation relies on lots of branching, and benefits
from a microbenchmark setup where it can hog the entire branch predictor. The
new variants only branch on the "is small" and "is overflow" checks, and are
otherwise straightline code.

Reviewed By: iahs

Differential Revision: D37074342

fbshipit-source-id: 6975fe07aa732de47c8ed26c6ed4b46ef43c281e