Add support for load shedding via token bucket (QPS ratelimiting)
Summary:
This is a counterpart to active request load shedding. Both can be applied and have separate but overlapping use cases.
The reason we are adding this is to have it as an option for CPUConcurrencyController; the use case we care about is avoiding over-eager load shedding due to downstreams getting slow, which we have seen in production already causing a few issues. Concurrency limits are particularly susceptible to this.

Note: One of the downsides of this approach is it is not really useful for memory overload protection, as we don't prevent build up of internal queues due to holding active requests. For example, if one of your backends gets so slow that you accumulate too many buffered requests.

We will consider these cases separately.

Reviewed By: roddym

Differential Revision: D37371231

fbshipit-source-id: 0f06ffa2e0946beea3ec2a76642057501e1d6d84