allow delayed resolution of built-in patch types
Summary:
Without this, targets like //icsp/apiserver/if:apiserver-cpp2-types fail
to build.  They only appear to fail for the _generated_ portion of the
built-in patch types (e.g., OptionalStringPatch).

I'm not quite sure why this is, but I suspect that it has to do with
the order in which mutators are run.

Reviewed By: Alfus, joshkehn

Differential Revision: D37705550

fbshipit-source-id: dd05ba0e23ad14408e78efc2bdf3c81a1d4d6695