Add support for patching string/binary protocol values
Summary:
This change is a bit tricky. While patching strings and binaries is defined as separate types of patches, the protocol::Value doesn't really distinguish between them when parsing serialized data. Meaning that if value is obtained via deserialization there is no way to distinguish them. Hence applying patch is implemented only for Binary type and string is just converted back and forth.

This is not ideal because e.g. appending to binary is not possible right now.

Reviewed By: Alfus

Differential Revision: D37658416

fbshipit-source-id: 0bb80c4ab3f283a9931676de75e895de566d49e3