Add support for setting QPS limits via CPUConcurrencyController
Summary:
Concurrency limits have two main issues when used with concurrency controller:

(1) They are hard to estimate at future load values (active requests is a gauge/instantaneous metric).
(2) It is an over-eager shedding mechanism and does not always correlate with CPU load. This can partially be offset as the algorithm will increase if we are under CPU usage, but it is still undesirable as it can affect prod and there are also cases which aren't properly mitigated.

This diff adds support for running the exact same algorithm, but instead it is tuning the token bucket limiter via `ThriftServer::setMaxQps()`.

Reviewed By: roddym

Differential Revision: D37746657

fbshipit-source-id: fd77455c3be75d3c6e5703f344c717c6b6a17951