Allow gen_struct_dependency_graph to handle fields whose types don't actually resolve
Summary:
The `gen_struct_dependency_graph` algorithm presumes that the types of all structures in the dependency graph actually resolve to something.  This might not be the case if this algorithm were run on a malformed file, as might be the case when linting a file.

This updates the dependency graph algorithm to attempt to resolve types, then to discard any type that does not resolve.  Such types will thus be omitted from the dependency graph.

Reviewed By: yfeldblum

Differential Revision: D37470666

fbshipit-source-id: cf7ca38e2e87280c1f52cf2135f8a761eb3d99c9