Support customizing serializedSize[ZC] on adapted types
Summary:
We currently determine the serialized size of an adapted type by converting it to its thrift type and calling serializedSize on that. This information is used to size rpc buffers into which serialization is then performed. As a result we call toThrift twice on the same value, which is a problem for types with expensive conversions or complex lifetime models (e.g. unique_ptr passed over RPC in-process). To solve this we let the adapter tell the runtime how large the serialized value will be in order to prevent this double conversion.
We only attempt to support this for fields with one total adapter. Fields with both field and type adapters would need both adapters to communicate to figure out the serialized size, which seems too complicated. We could support letting the outermost adapter do this as well, but that is not done here because it seems quite niche and increases codegen complexity.

Reviewed By: thedavekwon

Differential Revision: D37836776

fbshipit-source-id: 49abb9a25d1dfb38e8c252e1c89f2cff58285f56