Avoid calling toThrift in serializedSize when the standard type has size invariant in its content
Summary:
All protocols provide constant values for serializedSize[ZC] of numeric types to avoid paying encoding cost twice. We can leverage this to reduce the instances where adapters have to provide their own implementation to avoid double-conversion, though this makes the contract slightly more complicated.
This is mostly relevant for unique_ptr adapted types, which are both the only instance of move-only types we've seen and are serialized to integers.

Reviewed By: yfeldblum

Differential Revision: D37967073

fbshipit-source-id: 3597207fd857a0885afd78bfb45af8af093ea3cb