fix broken deepcopy on thrift-python structs
Summary:
The `deepcopy` operator is not implemented for thrift-python structs, which causes an error when attempting to deepcopy a thrift struct or any object (list, dict, etc) with a thrift struct in it.

It turns out that the copy operator is actually just the deepcopy operator, so just implement the deepcopy operator with the copy operator.

Reviewed By: nanshu

Differential Revision: D38123062

fbshipit-source-id: ced55c90051c11b0248d183545f4486552d93a30