Add op::ensure
Summary:
Added `op::ensure(field)` method. This emplaces the field if it doesn't exist, and it works for both field_refs and smart pointers. For field_refs, it calls ensure on the field, and for smart pointers, it constructs a new smart pointer if it doesn't exist.

Also added `ensureValue` method in `BasePatch.h` which works for field refs (same API as clearValue in the same file).

It doesn't support adapted field, smart pointers with custom allocators, and union. They will be supported in the next diff.

Reviewed By: Alfus

Differential Revision: D37947758

fbshipit-source-id: 7eaec30cdfdf8e9edac83df2dba4a369b0963988