Add op::copy
Summary:
Added `op::copy(src_field, dst_field)` method. This copies the field from src to dst, and it works for both field_refs and smart pointers. For field_refs, it simply calls copy_from on method. For unique pointers, if src is nullptr, it sets dst to nullptr, otherwise, it constructs a new unique pointer with the same data if it doesn't exist. For shared pointers, it simply shares the ptr. This is consistent with how thrift struct implements assignment operator for smart pointer fields.

It doesn't support adapted field, smart pointers with custom allocators, and union. They will be supported in a different diff.

Reviewed By: thedavekwon

Differential Revision: D38063944

fbshipit-source-id: c82417b7088c0e25241c3c2cf327f309a3f6d0e2