Fix py3 interaction test
Summary: Whoops forgot an await in the first test I wrote then copy pasted that missing await into every test. The tests still work since I guess the request itself is so lightweight and it's done on localhost, so it completes in time before the next request happens. Still this is wrong and these statements should be awaited.

Reviewed By: junzh0u

Differential Revision: D38410761

fbshipit-source-id: e12b3d111a17c9f7e92bf57dc70ea2c8ff48aaab