Improve handling of mstch element positions
Summary:
Scientists discovered that an mstch object is either a part of some container or not. There is no partial containment. With this deep insight merge `ELEMENT_POSITION` and element index into `mstch_element_position` so that element index and first/last indicators are always passed together.

This simplifies Thrift mstch objects, makes the factory API more regular and removes the need for `basic_mstch_factory_impl` (now removed). As a side effect this fixes a latent bug in `mstch_const_value::const_struct` which used a field ID as an index. Fortunately the latter wasn't used anywhere (yet).

Reviewed By: thedavekwon

Differential Revision: D38426517

fbshipit-source-id: 852ca84b48d1ec3244f1a65948caee8193374348