improve container element access/iteration efficiency
Summary:
Currently, accessing container element requires unpacking the element value from internal format (e.g. a tuple for a struct) to a python format(e.g. a Struct object for a struct). The unpacked object is not cached so subsequent accesses of the same element will incur another unpacking/conversion. This is not ideal for a few use cases. E.g. a user case may cache a thrift-python struct object somewhere, and iterate one of its list fields (which stores a long list of another struct) in a very hot call path, and every time it needs to create a new Struct from a tuple for each element it reads from the list. Similarly for a list of str, since we store bytes as str's internal format, a `bytes.decode()` was called for every str element access.

In this diff, we make container `_fbthrift_data` store "Python" objects (e.g. Struct for struct, Str for string) instead of internal format objects (e.g. tuple for struct, bytes for string), so subsequent calls to access/iterate these containers won't incur conversions. The same Python object is returned.

Reviewed By: junzh0u

Differential Revision: D38403971

fbshipit-source-id: e0714e8c48472a928dd83557b2bcc5628b81e169