Remove broken program name override in generator
Summary:
`t_generator` has a program name override mechanism in the form of a virtual function `get_program_name`. However, it doesn't work because this function is called in a constructor and, quoting https://isocpp.org/wiki/faq/strange-inheritance#calling-virtuals-from-ctors:

> In a constructor, the virtual call mechanism is disabled because overriding from derived classes hasn’t yet happened.

Neither it is used or needed so remove it in favor of getting the name from the program directly.

(Note: this ignores all push blocking failures!)

Reviewed By: thedavekwon

Differential Revision: D38521435

fbshipit-source-id: e39e772bbce558f1cb130010d776bf9dfd7a850d