Correct error type in stream elements
Summary:
In the current code, the function signature looks like this:

```
async fn method_name()
    -> Result<
        BoxStream<'static, Result<StreamElementType, MethodNameStreamError>>,
        MethodNameExn,
    >;
```

This isn't correct because `MethodNameStreamError` (and `MethodNameError`) are supposed to be used on the client side. They usually look like this:

```
enum MethodNameError {
    ex(DeclaredException),
    ApplicationException(::fbthrift::ApplicationException),
    ThriftError(::anyhow::Error),
}
```

Note that `ThriftError` is used to cover things like `TServiceRouterExceptions` which are not produced on the server side.

On the server side, we should use `MethodNameExn` and `MethodNameStreamExn` which does not have the `ThriftError` variant.

Reviewed By: Imxset21

Differential Revision: D38674207

fbshipit-source-id: 0edcc33d282a4e6d56edd8a7a837b8d42e437374