Added unit test for TProtocol
Summary:
ByteBufTJSONProtocol was throwing null pointer exceptions when used as serialization protocol.
The NPE happened in the `trans_.write` calls as the TTransport is null and should not be used on ByteBuffer protocols (since the data should be always written to the wapped ByteBuffer only).

**Fix:** replaced calls to `trans_.write` with `getByteBuf().writeByte`.

There was another issue when parsing the JSON too where some bytes were "skipped". The problem happened because there was a mix of calls to LookaheadReader to peek bytes with direct subsequent calls to the wrapped ByteBuffer to read bytes, which would then skip the previously peeked byte.

**Fix:** Replaced all direct calls to `getByteBuf().readByte()` with `reader_.read()`.

There was one more issue in the `getTypeIDForTypeName` function where it was reading a byte to determine if there were more bytes to read and consequently skipping that byte.

**Fix:** Replaced that with `name.readableBytes() > 0`

Differential Revision: D35717837

fbshipit-source-id: 88f6746f464e2146b0af138325c7f79635f9cd99