update thread pool settings
Summary: Changes thread pool settings so that the default number of off loop threads is equal to the number of eventloop threads. Lowers the number of event loop threads to equal the number of CPU codes. Adds a configuration to set the number of pending tasks before creating a new thread. Lowers the default number of pending tasks before creating a new thread to reactor-core queues.xs_buffer_size small.

Reviewed By: avalonalex

Differential Revision: D38730357

fbshipit-source-id: f028705961ef9961b9a5f2741e4316dad44d9287