Implement extractMaskFromPatch without nested masks
Summary:
Implemented `extractMaskFromPatch` method which constructs a Mask that only contains fields that are modified by the Patch. Constructing nested Mask for map and object patches is in next diff.
```
Mask extractMaskFromPatch(const protocol::Object& patch);
```
For map, it uses the address of Value key as the key for the mask.
Note that Mask contains pointer to `protocol::Value` in patch, so caller needs to make sure Patch has longer lifetime than the mask.

Differential Revision: D38673551

fbshipit-source-id: cce9e2023e1899ade0da65b448fc257d7aae3c56