Fix source range computation for statements
Summary:
Trailing `,` or `;` syntactically belongs to a statement so include it in statement's source range. Apart from theoretical correctness considerations this will be important for codemods because otherwise removing the text corresponding to a statement's source range may result in an invalid program containing a dangling `,` or `;`.

Also add range tracker to make source range computation easier.

Reviewed By: iahs

Differential Revision: D38791993

fbshipit-source-id: 67b8928c18e243ed19cd0c210da238255fa80135