Avoid coroutine code generation in coroutine service handler stubs
Summary:
**Better coroutine build times**!

Code generation for coroutines with clang is very slow. This has been the big blocker in landing `co_*` method stubs by default (and obsoleting `cpp.coroutine`).

By coroutine codegen, I am specifically referring to use of `co_await`, `co_yield`, and `co_return`. The current method stubs use `folly::coro::toTask` which uses coroutines underneath the hood.

Thrift server has a cascading mechanism at runtime to determine which C++ method the user-provided handler has implemented:
`async_tm_*` → `co_*` → `future_*` → `semifuture_*` → `<unprefixed>`.
The nodes on the left of the arrow wrap the method to their right, and so on.

This diff provides a default implementation for `co_*` which no longer wraps `future_*` but still maintains the cascading logic. It throws an `UnimplementedCoroMethod` exception which is caught by `async_tm_*` - this is a signal for `async_tm_*` to move on and try `future_*`.

This is huge - we avoid using `folly::coro::toTask` and thus avoid coroutine codegen. This should improve build times by several times.

One caveat of this approach is that `co_*` *consume* the arguments they are provided. i.e. `std::unique_ptr` objects are moved in when calling it. Because we don't want to change the public interface of `co_*`, the challenge was figuring out how to pass these arguments back to `async_tm_*`. The solution here is to move the arguments (type-erased) into `UnimplementedCoroMethod` and use the native ABI for exceptions carry it back to the caller.

All of this machinery is not great for performance. However, this only happens once per method - the cascading mechanism "remembers" which method call was implemented by the user and in a future invocation can have `async_tm_*` directly invoke that method (e.g. `async_tm_*` → `semifuture_*`). So performance is not a concern.
 ---

Reviewed By: iahs

Differential Revision: D38590146

fbshipit-source-id: 176f149f8066a6ac2b807aa75d081c1754487c98