Unhack generator option handling
Summary:
`t_generation_context` consists of generator output options and a source manager with the manager added there only to avoid changing numerous generator ctors. Now that the latter is not a problem add source manager as a parameter to the generator ctor where it should be, move output options to `process_options` and remove `t_generation_context`. Apply the template method pattern to `process_options` to guarantee that some options are always handled by the base generator and processing of the rest can be overridden by subclasses.

Remove `thrift/compiler/test/t_mstch_generator_test.cc` which doesn't add a lot of value.

And finally rename (the inverse of) `is_out_path_absolute` to `add_gen_dir` because this option has nothing to do with absolute paths. It controls whether to add `gen-<name>` directory to the output directory.

Reviewed By: avalonalex

Differential Revision: D38812260

fbshipit-source-id: 07070a3b9122fd53a4b59408fa8f1aa5c4ebc6b5