Remove dupe op from spec and add missing code gen fields
Summary:
On the wire set/maps/structs are sequences of key-value pairs, while lists/string/binary are sequences of values. Which means, that 'add' is 'add to beginning of a sequence' and 'put' is 'put at the end of a sequence' for all container/sequence types, so we don't need a dedicated 'prepend' op.

I also cleaned up the documentation and code a bit, added missing fields from the updated spec, with a todo to remove optional patch usage completely. The c++ wrappers remain unchanged.

(Note: this ignores all push blocking failures!)

Differential Revision: D38819785

fbshipit-source-id: 8cdeaf03a3be94c0dbe7528f2833d2a07ec0f5cb