Add source locations to doc comments
Summary: Add source locations to documentation comments and stop removing empty comments. We shouldn't remove empty comments because AST should be a faithful representation of the input and it's not a useful optimization anyway. In particular this information can be useful for tooling to give some diagnostic on useless documentation comments.

Reviewed By: avalonalex

Differential Revision: D38835885

fbshipit-source-id: 656ae839019bde9447d347859ffe8d38bf56a5b1