Implement extractMaskFromPatch for nested masks
Summary:
Implemented extractMaskFromPatch method which constructs a Mask that only contains fields that are modified by the Patch. It constructs nested Mask for map and object patches
```
Mask extractMaskFromPatch(const protocol::Object& patch);
```
For map, it uses the address of Value key as the key for the mask.
Note that Mask contains pointer to protocol::Value in patch, so caller needs to make sure Patch has longer lifetime than the mask.

Reviewed By: Mizuchi

Differential Revision: D38725369

fbshipit-source-id: 7b24269660da3c32342eb0d8fa6e72a89a280c5e