Add thrift service name to MethodMetadata
Summary:
So that Trusted API Enforcer is able to get the config for the thrift service and decide if we need to enforce trusted calls or not. e.g.
https://www.internalfb.com/code/configerator/source/privacy/constraints/enforcer/thrift/BlobStore.cconf

Also the thrift service name is used for PZ check in WWW:
https://www.internalfb.com/code/www/[3e04754ab025]/flib/purpose/cipp/handlers/TClientPoliciedAsyncHandler.php?lines=83-87

We will add a new RouteHandle for TAE and PZ check. At first we will do this at the thrift service level just like what we do in WWW. Later we may introduce the asset level enforcement for data systems.

Reviewed By: vitaut

Differential Revision: D38780214

fbshipit-source-id: 6a058ac678ffd58a8fe1c78b3996a2f0c2fc5e54