Fixing documentation warnings for thrift
Summary:
When using thrift in iOS, multiple compiler flags need to be set in order to suppress warnings as errors. One of these flags is
```"-Wno-documentation"```
to suppress warnings as errors that come from documentation mistakes in the xplat/thrift directory. From inspecting the thrift library, the comments that the iOS compiler points out as false are obviously wrong, referencing variable names that no longer are used in the function.

Changed the incorrect comments to match up with the current variable names in the functions removing the need to include "-Wno-documentation" in fbobjc BUCK compiler flags.

Reviewed By: natestedman

Differential Revision: D39062878

fbshipit-source-id: fb8ee4de04d3d36523a7707ea61ce420e030fcc5