fix extractMaskFromPatch
Summary: `extractMaskFromPatch` was returning allMask when one of the patch op is not an intrinsic default and has either clear, add, or put field. This doesn't handle the case when those fields are false and it has fieldPatch. This diff fixes this issue by checking intrinsic default for those fields.

Differential Revision: D39220955

fbshipit-source-id: adc923ea05c21e0441ba60e543666320d9c94feb