TerseWrite Struct Field Optimization
Summary: A struct is empty iff all fields are optional or terse, and if all optional fields are unset and terse field is equal to the intrinsic default. We currently have a naive implementation that already does this, but we need an optimization since there is a worst case N^2 field empty check when there is N-level of terse struct field.

Reviewed By: Mizuchi

Differential Revision: D39179261

fbshipit-source-id: e2184857d40d860a4d72a7f38fdaaf51f946b6e5