Make interaction method calls work
Summary:
Interaction method calls require the correct method metadata to be passed to the transport layer, so doing that here.

I'm specifying these parameters in the python layer since that's where we have codegen, then passing them to the cython library which creates a metadata object to wrap up these params and pass them over to the cpp client library.

Reviewed By: ParalaxRus

Differential Revision: D38605761

fbshipit-source-id: e5b76ea824a37974d23d173e5bf78ec683d237f7