Implement metadata response in thrift python server
Summary:
This method is required to allow any sort of metadata query of a thrift python server to get the methods it supports.

Unfortunately it expects a thrift cpp2 object, so the best solution here is to just serialize and deserialize a thrift python object, which carries a performance cost, but is acceptable given that this method is primarily used for debugging and testing.

Reviewed By: junzh0u

Differential Revision: D38956622

fbshipit-source-id: b9ec4b4a87bc7ae3c6bb995bcb682c4d11cd079c