Add logging around whether resource pools are enabled
Summary: In case of Sidecar, as many as 3 thrift servers are started in the same binary. To make it easier to make sense of which one has ResourcePools enabled and which doesn't this diff adds port / socket path to a log line to differentiate between thrift servers within the same process.

Reviewed By: roddym

Differential Revision: D39476970

fbshipit-source-id: f4668c996bcaf280280993bc176c6edced289c39