Remove setThreadManager from axon
Summary: We are deprecating the thrift ThreadManager - it will be replaces with resource pools which use a regular folly executor. As part of this we are replacing code that directly sets the thread manager with code that specifies the requirements for its thread management. At the moment this will continue to create a ThreadManager, just inside thrift server instead of in client code. Longer term, the thrift server will switch to creating the right resource pools setup when the service is migrated.

Reviewed By: al13n321

Differential Revision: D39108418

fbshipit-source-id: 41f26d36a824f604aa097c5c0185aa82e88f60a1