Add AttributeSource::OVERRIDE_INTERNAL
Summary:
`AttributeSource::OVERRIDE_INTERNAL` is a new, highest priority AttributeSource for internal (to Thrift) use only.

To enable this change, the AttributeSource type has been changed from a scoped enum to a class with an enum-like interface. This also allows the new `OVERRIDE_INTERNAL` value to only be accessible by future friended classes.

Reviewed By: praihan, joshkirstein

Differential Revision: D39936318

fbshipit-source-id: 51e3fed695c612ad5de2f1720bd19b9e3fe00db0